[API] create check sco account recovery service
